requirements.txt
google-generativeai
openai
pymupdf
Pillow
pydub
edge-tts
nest_asyncio
python-dotenv